miner/requirements.txt
elasticsearch_dsl==5.1.0
